requirements.txt
flask==3.0.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
